requirements.txt
Django==5.0.2
requests==2.31.0
python-dotenv==1.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1 